add render to api
