+++ api/handler/trigger.go
@@ -3,18 +3,16 @@ package handler
 import (
 	"fmt"
 	"net/http"
-	"time"
 
 	"github.com/go-chi/chi"
 	"github.com/go-chi/render"
-	"github.com/go-graphite/carbonapi/date"
-	"github.com/moira-alert/moira/remote"
 
 	"github.com/moira-alert/moira/api"
 	"github.com/moira-alert/moira/api/controller"
 	"github.com/moira-alert/moira/api/dto"
 	"github.com/moira-alert/moira/api/middleware"
 	"github.com/moira-alert/moira/expression"
+	"github.com/moira-alert/moira/remote"
 	"github.com/moira-alert/moira/target"
 )
 
@@ -28,11 +26,7 @@ func trigger(router chi.Router) {
 		router.Get("/", getTriggerThrottling)
 		router.Delete("/", deleteThrottling)
 	})
-	router.Route("/metrics", func(router chi.Router) {
-		router.With(middleware.DateRange("-10minutes", "now")).Get("/", getTriggerMetrics)
-		router.Delete("/", deleteTriggerMetric)
-		router.Delete("/nodata", deleteTriggerNodataMetrics)
-	})
+	router.Route("/metrics", triggerMetrics)
 	router.Put("/maintenance", setMetricsMaintenance)
 }
 
@@ -117,45 +111,6 @@ func deleteThrottling(writer http.ResponseWriter, request *http.Request) {
 	triggerID := middleware.GetTriggerID(request)
 	err := controller.DeleteTriggerThrottling(database, triggerID)
 	if err != nil {
-		render.Render(writer, request, err)
-	}
-}
-
-func getTriggerMetrics(writer http.ResponseWriter, request *http.Request) {
-	triggerID := middleware.GetTriggerID(request)
-	fromStr := middleware.GetFromStr(request)
-	toStr := middleware.GetToStr(request)
-	from := date.DateParamToEpoch(fromStr, "UTC", 0, time.UTC)
-	if from == 0 {
-		render.Render(writer, request, api.ErrorInvalidRequest(fmt.Errorf("can not parse from: %s", fromStr)))
-		return
-	}
-	to := date.DateParamToEpoch(toStr, "UTC", 0, time.UTC)
-	if to == 0 {
-		render.Render(writer, request, api.ErrorInvalidRequest(fmt.Errorf("can not parse to: %v", to)))
-		return
-	}
-	triggerMetrics, err := controller.GetTriggerMetrics(database, from, to, triggerID)
-	if err != nil {
-		render.Render(writer, request, err)
-		return
-	}
-	if err := render.Render(writer, request, &triggerMetrics); err != nil {
-		render.Render(writer, request, api.ErrorRender(err))
-	}
-}
-
-func deleteTriggerMetric(writer http.ResponseWriter, request *http.Request) {
-	triggerID := middleware.GetTriggerID(request)
-	metricName := request.URL.Query().Get("name")
-	if err := controller.DeleteTriggerMetric(database, metricName, triggerID); err != nil {
-		render.Render(writer, request, err)
-	}
-}
-
-func deleteTriggerNodataMetrics(writer http.ResponseWriter, request *http.Request) {
-	triggerID := middleware.GetTriggerID(request)
-	if err := controller.DeleteTriggerNodataMetrics(database, triggerID); err != nil {
 		render.Render(writer, request, err)
 	}
 }
