+++ checker/timeseries_test.go
@@ -112,7 +112,7 @@ func TestGetTimeSeries(t *testing.T) {
 				}},
 				Wildcard: true,
 			}
-			expected := &triggerTimeSeries{
+			expected := &TriggerTimeSeries{
 				Main:       []*target.TimeSeries{&timeSeries},
 				Additional: make([]*target.TimeSeries, 0),
 			}
@@ -135,7 +135,7 @@ func TestGetTimeSeries(t *testing.T) {
 				StepTime:  retention,
 				Values:    []float64{0, 1, 2, 3, 4},
 			}
-			expected := &triggerTimeSeries{
+			expected := &TriggerTimeSeries{
 				Main:       []*target.TimeSeries{{MetricData: types.MetricData{FetchResponse: fetchResponse}}},
 				Additional: make([]*target.TimeSeries, 0),
 			}
@@ -167,7 +167,7 @@ func TestGetTimeSeries(t *testing.T) {
 			}
 			addFetchResponse := fetchResponse
 			addFetchResponse.Name = addMetric
-			expected := &triggerTimeSeries{
+			expected := &TriggerTimeSeries{
 				Main:       []*target.TimeSeries{{MetricData: types.MetricData{FetchResponse: fetchResponse}}},
 				Additional: []*target.TimeSeries{{MetricData: types.MetricData{FetchResponse: addFetchResponse}}},
 			}
@@ -232,7 +232,7 @@ func TestGetTimeSeries(t *testing.T) {
 }
 
 func TestGetTargetName(t *testing.T) {
-	tts := triggerTimeSeries{}
+	tts := TriggerTimeSeries{}
 
 	Convey("GetMainTargetName", t, func() {
 		So(tts.getMainTargetName(), ShouldResemble, "t1")
@@ -257,7 +257,7 @@ func TestGetExpressionValues(t *testing.T) {
 		timeSeries := target.TimeSeries{
 			MetricData: types.MetricData{FetchResponse: fetchResponse},
 		}
-		tts := &triggerTimeSeries{
+		tts := &TriggerTimeSeries{
 			Main: []*target.TimeSeries{&timeSeries},
 		}
 		expectedExpressionValues := &expression.TriggerExpression{
@@ -307,7 +307,7 @@ func TestGetExpressionValues(t *testing.T) {
 		timeSeriesAdd := target.TimeSeries{
 			MetricData: types.MetricData{FetchResponse: fetchResponseAdd},
 		}
-		tts := &triggerTimeSeries{
+		tts := &TriggerTimeSeries{
 			Main:       []*target.TimeSeries{&timeSeries},
 			Additional: []*target.TimeSeries{&timeSeriesAdd},
 		}
@@ -343,36 +343,36 @@ func TestGetExpressionValues(t *testing.T) {
 
 func TestTriggerTimeSeriesHasOnlyWildcards(t *testing.T) {
 	Convey("Main timeseries has wildcards only", t, func() {
-		tts := triggerTimeSeries{
+		tts := TriggerTimeSeries{
 			Main: []*target.TimeSeries{{Wildcard: true}},
 		}
 		So(tts.hasOnlyWildcards(), ShouldBeTrue)
 
-		tts1 := triggerTimeSeries{
+		tts1 := TriggerTimeSeries{
 			Main: []*target.TimeSeries{{Wildcard: true}, {Wildcard: true}},
 		}
 		So(tts1.hasOnlyWildcards(), ShouldBeTrue)
 	})
 
 	Convey("Main timeseries has not only wildcards", t, func() {
-		tts := triggerTimeSeries{
+		tts := TriggerTimeSeries{
 			Main: []*target.TimeSeries{{Wildcard: false}},
 		}
 		So(tts.hasOnlyWildcards(), ShouldBeFalse)
 
-		tts1 := triggerTimeSeries{
+		tts1 := TriggerTimeSeries{
 			Main: []*target.TimeSeries{{Wildcard: false}, {Wildcard: true}},
 		}
 		So(tts1.hasOnlyWildcards(), ShouldBeFalse)
 
-		tts2 := triggerTimeSeries{
+		tts2 := TriggerTimeSeries{
 			Main: []*target.TimeSeries{{Wildcard: false}, {Wildcard: false}},
 		}
 		So(tts2.hasOnlyWildcards(), ShouldBeFalse)
 	})
 
 	Convey("Additional timeseries has wildcards but Main not", t, func() {
-		tts := triggerTimeSeries{
+		tts := TriggerTimeSeries{
 			Main:       []*target.TimeSeries{{Wildcard: false}},
 			Additional: []*target.TimeSeries{{Wildcard: true}, {Wildcard: true}},
 		}
